3d viewer changes
