+++ .osparc/3d-viewer-gpu/metadata.yml
@@ -1,10 +1,10 @@
-name: 3d-viewer
+name: 3d-viewer-gpu
 thumbnail: 
 description: "Paraview Web-based Visualizer"
-key: simcore/services/dynamic/3d-viewer
-version: 2.11.1
+key: simcore/services/dynamic/3d-viewer-gpu
+version: 3.0.5
 integration-version: 2.0.0
-version_display: 2.11.1
+version_display: 3.0.5
 type: dynamic
 authors:
   - name: "Sylvain Anderegg"
