+++ .osparc/3d-viewer-gpu/metadata.yml.j2
