+++ services/3d-viewer-gpu/.cookiecutterrc
