+++ services/3d-viewer-gpu/.dockerignore
