+++ services/3d-viewer-gpu/VERSION
