+++ services/3d-viewer-gpu/docker/custom/entrypoint.bash
