+++ services/3d-viewer-gpu/docker/custom/healthcheck_curl_host.py
