+++ services/3d-viewer-gpu/metadata/metadata-gpu.yml
