+++ services/3d-viewer-gpu/src/3d-viewer/config/s4lColorMap.json
