+++ services/3d-viewer-gpu/src/3d-viewer/config/visualizer_config.json
