+++ services/3d-viewer-gpu/src/3d-viewer/rebrand-osparc/rebrand-osparc.patch
