+++ services/3d-viewer-gpu/src/3d-viewer/rebrand-osparc/to-be-copied/src/panels/ProgressLoaderWidgetOsparc/ProgressLoaderWidgetOsparc.mcss
