+++ services/3d-viewer-gpu/tests/integration/test_docker_image.py
