+++ services/3d-viewer-gpu/tools/update_compose_labels.py
