+++ services/3d-viewer-gpu/versioning/integration.cfg
